+++ src/components/MarkdownHighlighter/MarkdownHighlighter.tsx
@@ -13,10 +13,11 @@ export default function MarkdownHighlighter({ markdown }: MarkProps) {
 
     return (
         <div className='my-2 mx-6'>
-            <div className="max-w-none prose prose-pre:m-0 prose-pre:p-0 prose-sm dark:prose-invert prose-hr:prose-invert prose-hr:my-6">
+            <div className="">
                 <Markdown
                     remarkPlugins={[remarkGfm]}
                     rehypePlugins={[rehypeRaw]}
+                    
                     components={{
                         code({ node, className, children: codeChildren, ...props }) {
                             const match = /language-(\w+)/.exec(className || "");
@@ -32,21 +33,32 @@ export default function MarkdownHighlighter({ markdown }: MarkProps) {
                                     {String(codeChildren).replace(/\n$/, "")}
                                 </SyntaxHighlighter>
                             ) : (
-                                <code className={className} {...props} > 
+                                <code className={className} {...props} >
                                     {codeChildren}
                                 </code>
                             );
                         },
 
                         p({ node, children, ...props }) {
-                          return (
-                            <p
-                              style={{ marginBottom: "0.5rem", whiteSpace: "pre-line" }}
-                              {...props}
-                            >
-                              {children}
-                            </p>
-                          );
+                            return (
+                                <p
+                                    style={{ marginBottom: "0.5rem", whiteSpace: "pre-line" }}
+                                    {...props}
+                                >
+                                    {children}
+                                </p>
+                            );
+                        },
+
+                        a({ node, children, ...props }) {
+                            return (
+                                <a
+                                    target='_blank'
+                                    {...props}
+                                >
+                                    {children}
+                                </a>
+                            );
                         },
                     }}
                 >